refactor(phase0): move token-engine tests under tests/unit/ (match layout)
tests/ is split by test TYPE (unit/integration/e2e/relay), and tests/unit/ is
organized by module (core, modules, serialization, transport, ...). The
token-engine unit tests now live at tests/unit/token-engine/ to follow that
convention instead of a top-level tests/token-engine/ folder. Import depth
adjusted; 23 tests still green.
